% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hips-package.R
\docType{data}
\name{adl_test}
\alias{adl_test}
\title{Adelaide test cohort scalars and inference results}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4568 rows and 11 columns.}
\usage{
adl_test
}
\description{
data.frame with 4,568 examples and 11 variables
}
\seealso{
\code{\link{adlCohort}}

Other cohorts: \code{\link{adlCohort}},
  \code{\link{caseControlCohorts}},
  \code{\link{hipsCohort_}}, \code{\link{scalars}}
}
\concept{cohorts}
\keyword{datasets}
