% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_df.R
\name{adlCohort}
\alias{adlCohort}
\title{Fetch Adelaide Test Samples split into differet subsampling cohorts}
\usage{
adlCohort(binarized = FALSE)
}
\arguments{
\item{binarized}{lgl(1) whether to binarize all features (eg, for OR computation)}
}
\value{
list of cohort scalar tables
}
\description{
Fetch Adelaide Test Samples split into differet subsampling cohorts
}
\examples{
adlCohort()
adlCohort(binarized=TRUE)
}
\seealso{
Other cohorts: \code{\link{adl_test}},
  \code{\link{caseControlCohorts}},
  \code{\link{hipsCohort_}}, \code{\link{scalars}}
}
\concept{cohorts}
