% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComplexSummary.R
\name{BaseSummary}
\alias{BaseSummary}
\title{Compute hip-cohort-table-specific individual statistics}
\usage{
BaseSummary(x, grp = character(0), ...)
}
\arguments{
\item{x}{data.frame, hip cohort table}

\item{grp}{x_chr, names of grouping columns}

\item{...}{additional stats to compute}
}
\description{
Default stats hardcoded into \code{BaseSummary}, additional stats
can be passed to \code{\link[dplyr]{summarise}} via \code{...}
}
\examples{
grp_lst <- list(character(0), "sex", "fx", c("sex", "dept"))
map(grp_lst, BaseSummary, x=hips::scalars)
map(grp_lst, BaseSummary, x=hips::scalars, bmi_mean = mean_(bmi))
BaseSummary(hips::scalars \%>\% filter(fx), grp = c("sex", "device_model"))
}
\seealso{
ComplexSummary
}
