% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hips-package.R
\docType{data}
\name{caseControlCohorts}
\alias{caseControlCohorts}
\title{matched cohort enrollment lists}
\format{An object of class \code{list} of length 4.}
\usage{
caseControlCohorts
}
\description{
List of 3 chr(1566) containing ids for participants in each of 3 case-control cohorts
}
\seealso{
Other cohorts: \code{\link{adlCohort}},
  \code{\link{adl_test}}, \code{\link{hipsCohort_}},
  \code{\link{scalars}}
}
\concept{cohorts}
\keyword{datasets}
