% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_dfs_fct_match}
\alias{check_dfs_fct_match}
\title{Verify that all factor levels in df2 are available in df1}
\usage{
check_dfs_fct_match(df1, df2, .silent = FALSE)
}
\arguments{
\item{df1}{data.frame with 1+ factor variables}

\item{df2}{data.frame with 1+ factor variables}
}
\value{
invisible.  stops runtime if new inference levels are found in df2
}
\description{
Used to ensure a model trained on df2 can be applied to all examples in df2
}
\examples{
df2 = gss_cat
df2$race <- as.character(df2$race)
df3 = gss_cat \%>\% filter(race \%in\% c("Black", "White"))
check_dfs_fct_match(gss_cat, gss_cat)
try(check_dfs_fct_match(gss_cat, df2))
try(check_dfs_fct_match(gss_cat, df3))
}
\seealso{
Other checks: \code{\link{check_y}}
}
\concept{checks}
