% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_y}
\alias{check_y}
\title{Confirm that a target vector is suitable for a classification model.}
\usage{
check_y(y)
}
\arguments{
\item{y}{x_fct or x_chr to with true labels}

\item{modelType}{chr(1)}
}
\value{
silent; raises if ill-suited model target
}
\description{
copied from \code{\link[caret]{train}}.
}
\examples{
check_y(iris[, 5])
suppressMessages(check_y(iris[, 5]))
try(check_y(fct_expand(iris[, 5], "new_lvl")))
try(check_y(fct_recode(iris[, 5], "3x" = "setosa")))
}
\seealso{
Other checks: \code{\link{check_dfs_fct_match}}
}
\concept{checks}
