% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_roc.R
\docType{class}
\name{Classifier}
\alias{Classifier}
\title{Classifier Abstract Base Class.}
\description{
Classifier Abstract Base Class.
}
\section{Slots}{

\describe{
\item{\code{Y}}{logical or binary integer. the true labels.}

\item{\code{id}}{character. Image IDs.}
}}

\concept{classifiers}
