% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_roc.R
\name{compare_cCs}
\alias{compare_cCs}
\title{Compare 2 ROC curves for AUC equivalence}
\usage{
compare_cCs(cC_lst = list(), compare_test = "delong", pilot = FALSE)
}
\value{
h.test object
}
\description{
Compare 2 ROC curves for AUC equivalence
}
