% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComplexSummary.R
\name{ComplexSummary}
\alias{ComplexSummary}
\title{Wrapper for fxns that generates combined table statistics by grouping}
\usage{
ComplexSummary(...)
}
\description{
Wrapper for fxns that generates combined table statistics by grouping
}
\examples{
ComplexSummary(hipsCohort())
ComplexSummary(hipsCohort(), grp = "sex")
}
\seealso{
Other viewTable: \code{\link{Kable}}
}
\concept{viewTable}
