% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_aes.R
\name{facet_cnn}
\alias{facet_cnn}
\title{Facet by CNN}
\usage{
facet_cnn(facets = ~cnn, scales = "free", as.table = FALSE,
  labeller = ggplot2::as_labeller(AES_CNN_LABELS), ...)
}
\description{
Facet by CNN
}
\examples{
data.frame(cnn = rep(c("rnd", "pre"), 5), n=1:10) \%>\% ggplot(., aes(x=n)) + geom_histogram() + facet_cnn()
}
\seealso{
Other gg_aes: \code{\link{PrettyPredictors}},
  \code{\link{PrettyTargets}}, \code{\link{gg_x_target}},
  \code{\link{scale_color_cnn}},
  \code{\link{scale_color_cohort}},
  \code{\link{scale_color_device}},
  \code{\link{scale_color_fx}},
  \code{\link{scale_color_targetMode}},
  \code{\link{scale_color_target}},
  \code{\link{scale_color_view}},
  \code{\link{scale_fill_sex}},
  \code{\link{scale_predictor_}},
  \code{\link{scale_x_cohort}},
  \code{\link{scale_x_target}}
}
\concept{gg_aes}
