% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{Fp_ml_dir}
\alias{Fp_ml_dir}
\title{Access ML dir contents}
\usage{
Fp_ml_dir(...)
}
\description{
Access ML dir contents
}
\examples{
Fp_ml_dir() \%>\% dir.exists()
Fp_ml_dir("foo.txt")
Fp_ml_dir("models", "foo.txt")
}
