% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{FS_hipsDir}
\alias{FS_hipsDir}
\title{Configured map of absolute project dir paths by system.}
\usage{
FS_hipsDir()
}
\description{
This serves a similar ad hoc functionality to the aaa.R .onLoad() routine in ProjUtilsRads.
}
\examples{
FS_hipsDir()
}
