% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{get_df_fct_lvls}
\alias{get_df_fct_lvls}
\title{Get a list of levels for each factor in a data.frame}
\usage{
get_df_fct_lvls(df, xtabs = FALSE)
}
\arguments{
\item{df}{a data.frame with 1+ factor variables}
}
\value{
A list with length equal to the number of factor variables in df.
}
\description{
A list is created that contains a x_chr for each factor variable
containing all unique factor levels
}
\examples{
get_df_fct_lvls(iris, xtabs=TRUE)
get_df_fct_lvls(forcats::gss_cat, xtabs=TRUE)
}
