% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_views.R
\name{gg_pc}
\alias{gg_pc}
\title{View a PC plot with indicated color/alpha}
\usage{
gg_pc(DF, COL)
}
\description{
View a PC plot with indicated color/alpha
}
\examples{
df <- data.frame(PC1 = rnorm(100), PC2 = rnorm(100), FX = rep(c(FALSE, TRUE), 100/2))
gg_pc(df, FX) + hips::scale_color_fx()
}
\seealso{
Other viewFig: \code{\link{ggMissing}}
}
\concept{viewFig}
