% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_aes.R
\name{gg_x_target}
\alias{gg_x_target}
\title{ggplot wrapper to make an x scale with angled labels}
\usage{
gg_x_target(...)
}
\description{
ggplot wrapper to make an x scale with angled labels
}
\examples{
gg_x_target() \%>\% s
}
\seealso{
Other gg_aes: \code{\link{PrettyPredictors}},
  \code{\link{PrettyTargets}}, \code{\link{facet_cnn}},
  \code{\link{scale_color_cnn}},
  \code{\link{scale_color_cohort}},
  \code{\link{scale_color_device}},
  \code{\link{scale_color_fx}},
  \code{\link{scale_color_targetMode}},
  \code{\link{scale_color_target}},
  \code{\link{scale_color_view}},
  \code{\link{scale_fill_sex}},
  \code{\link{scale_predictor_}},
  \code{\link{scale_x_cohort}},
  \code{\link{scale_x_target}}
}
\concept{gg_aes}
