% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_views.R
\name{ggMissing}
\alias{ggMissing}
\title{View data.frame missing data proportion for each variable with missingness}
\usage{
ggMissing(x)
}
\description{
View data.frame missing data proportion for each variable with missingness
}
\examples{
forcats::gss_cat \%>\% ggMissing()
hipsCohort() \%>\% ggMissing
datasets::mtcars \%>\% ggMissing()
}
\seealso{
Other viewFig: \code{\link{gg_pc}}
}
\concept{viewFig}
