% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OddsRatios.R
\name{trans_sig}
\alias{trans_sig}
\alias{ggOddsRatios}
\alias{ggOddsRatios.OddsRatios}
\alias{ggOddsRatios.cohort_df}
\title{View odds ratios of pair-wise target associations.}
\usage{
trans_sig(x)

ggOddsRatios(x, ...)

\method{ggOddsRatios}{OddsRatios}(x, sig = TRUE)

\method{ggOddsRatios}{cohort_df}(x, target_chr = "fx", sig = TRUE)
}
\arguments{
\item{x}{data.frame with odds ratios to view}

\item{sig}{lgl(1) should asterixis be added to designate significance?}
}
\description{
View odds ratios of pair-wise target associations.
}
\examples{
trans_sig(1/rnorm(100))
ggOddsRatios(hipsCohort(binary))
ggOddsRatios(OddsRatios(hipsCohort(binary)))
ggOddsRatios(hipsCohort(binary), sig = FALSE)
ggOddsRatios(OddsRatios(hipsCohort(binary)), sig = FALSE)
ggOddsRatios(OddsRatios(hipsCohort(binary) \%>\% split(.$device_model)), sig = TRUE) + aes_device_col()
}
\seealso{
Other OddsRatios: \code{\link{OddsRatios}}
}
\concept{OddsRatios}
