% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_roc.R
\name{glance.ClassifierCurve}
\alias{glance.ClassifierCurve}
\title{View summary stats for a classifier point.}
\usage{
\method{glance}{ClassifierCurve}(x, ...)
}
\arguments{
\item{x}{a \code{\link{Classifier}}}

\item{...}{dots, needed to stay matched to the generic}
}
\description{
View summary stats for a classifier point.
}
\examples{
data(classifiers, package="AnalysisToolkit")
glance(classifiers$classifier_points[[1]])
}
\concept{tidy, glance}
