% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hips-package.R
\docType{package}
\name{hips}
\alias{hips}
\alias{hips-package}
\title{Hips.}
\description{
Project package to analyze hip radiographs with statistical learning models.
}
\section{Cohort Constructors}{


Constructor functions for loading cohorts
\code{\link{hipsCohort_}}
\code{\link{adlCohort}}
}

\section{Model Training}{


Partition cohort table
\code{\link{StratifiedPartition}}

Train model
\code{\link{trainGlm}}
}

\section{Model Evaluation}{


\code{\link{predict_pY}}
}

\section{Views}{


Tables: \code{\link{ComplexSummary}}
Figures: \code{\link{gg_pc}}
}

\section{Stats}{


Odds Ratios: \code{\link{OddsRatios}}
}

\author{
\strong{Maintainer}: Marcus Badgeley \email{mabadgeley@gmail.com}

}
