% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_df.R
\name{hipsCohort_}
\alias{hipsCohort_}
\alias{hipsCohort}
\alias{print.cohort_df}
\title{Load a cohort table}
\usage{
hipsCohort_(mutating = "none", filtering = "crossSectional")

hipsCohort(mutating, filtering)

\method{print}{cohort_df}(x, ...)
}
\arguments{
\item{mutating}{chr(1) describing the mutated version of the cohort frame}

\item{filtering}{chr(1) describing the set of patients to include}
}
\value{
img-major data.frame with scalars and btlncks
}
\description{
Load a cohort table
}
\examples{
hipsCohort_()
hipsCohort_(mutating = "binary")
hipsCohort_(mutating = "binary", filtering = "caseControl_matchDem")
hipsCohort()
hipsCohort(mutating = binary)
hipsCohort(mutating = complete)
hipsCohort(mutating = complete, filtering = caseControl_matchDem)
hipsCohort() \%>\% print
hipsCohort(binary) \%>\% print
hipsCohort() \%>\% select(-img)
}
\seealso{
Other cohorts: \code{\link{adlCohort}},
  \code{\link{adl_test}}, \code{\link{caseControlCohorts}},
  \code{\link{scalars}}
}
\concept{cohorts}
