% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{hipsInfo_}
\alias{hipsInfo_}
\alias{hipsInfo}
\title{Convenience fxn to access pkg configuration info in inst/info}
\usage{
hipsInfo_(filestem)

hipsInfo(filestem)
}
\arguments{
\item{filestem}{chr(1)}
}
\value{
rds file contents, generally a data.frame
}
\description{
hipsInfo_ does SE, and hipsInfo is a NSE wrapper.
}
\examples{
hipsInfo_()
hipsInfo_("targets")
hipsInfo()
hipsInfo(targets)
}
