% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComplexSummary.R
\name{Kable}
\alias{Kable}
\alias{Kable.cohort_df}
\title{Kable S3 generic for kable printing with nice defaults}
\usage{
Kable(x, ...)

\method{Kable}{cohort_df}(x, grp = character(0), ...)
}
\arguments{
\item{x}{object to render in a report}
}
\description{
Kable S3 generic for kable printing with nice defaults
}
\examples{
Kable(hipsCohort())
Kable(x = hipsCohort(), grp = "sex")
Kable(x = hipsCohort(), grp = c("sex", "dept"))
}
\seealso{
Other viewTable: \code{\link{ComplexSummary}}
}
\concept{viewTable}
