% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OddsRatios.R
\name{keep_binary}
\alias{keep_binary}
\title{Only keep variables with two unique non-missing levels}
\usage{
keep_binary(df)
}
\description{
Used for pair-wise Odds Ratio computation
}
\examples{
hipsCohort() \%>\% keep_binary()
hipsCohort() \%>\% split(.$device_model) \%>\% map(keep_binary)
}
