% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OddsRatios.R
\name{OddsRatios}
\alias{OddsRatios}
\alias{OddsRatios.data.frame}
\alias{OddsRatios.list}
\title{Compute pair-wise odds ratios between binary variables and a target}
\usage{
OddsRatios(x, ...)

\method{OddsRatios}{data.frame}(x, target_chr = "fx")

\method{OddsRatios}{list}(x, target_chr = "fx",
  grp_chr = "device_model")
}
\value{
a covariate-major data.frame with S3 class 'OddsRatios'
}
\description{
Compute pair-wise odds ratios between binary variables and a target
}
\examples{
ors <- OddsRatios(hipsCohort(binary), "fx")
ors2 <- hipsCohort(binary) \%>\% split(.$device_model) \%>\% OddsRatios(target_chr = "fx", grp_chr = "device_model")
}
\seealso{
Other OddsRatios: \code{\link{trans_sig}}
}
\concept{OddsRatios}
