% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eval.R
\name{predict_pY}
\alias{predict_pY}
\title{Predict probas for a classification models positive class}
\usage{
predict_pY(model, ...)
}
\arguments{
\item{model}{caret classification model}

\item{...}{possibly a new data frame with argname newdata}
}
\value{
numeric vector with length equal to nrow in newdata df or the training df
}
\description{
Positive class identity inferred from model$levels
}
\seealso{
Other evalTrain: \code{\link{train_eg_ids}},
  \code{\link{train_n_eg}}, \code{\link{train_terms}}
}
\concept{evalTrain}
