% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComplexSummary.R
\name{PrettyEach}
\alias{PrettyEach}
\title{Beautify individual table statistics}
\usage{
PrettyEach(x, ...)
}
\arguments{
\item{x}{data.frame, hip cohort table}

\item{...}{additional args to pass to \code{\link[base]{format}}}
}
\description{
Beautify individual table statistics
}
\examples{
grp_lst <- list(character(0), "sex", "fx", c("sex", "dept"))
map(grp_lst, compose(PrettyEach, BaseSummary), x = scalars)
map(grp_lst, compose(partial(PrettyEach, trim=TRUE), BaseSummary), x=scalars)
}
\seealso{
ComplexSummary
}
