% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiographs.R
\name{readNpy}
\alias{readNpy}
\alias{readImg}
\alias{showImg}
\title{Read a numpy image file from disk}
\usage{
readNpy(fp)

readImg(img_id, dir = hipsOpt(dir_img_proc))

showImg(img_id, dir = hipsOpt(dir_img_proc))
}
\arguments{
\item{fp}{chr(1) file path to numpy image}
}
\value{
array
}
\description{
Uses the \code{reticulate} package to create an R representation of the numpy module
Created for issue-60
}
\details{
\code{readNpy} takes a numpy file path

\code{readImg} takes and image id and infers the path

\code{showImg} is fxn composition of \code{readImg} and \code{vizR::Viz}
}
\examples{
readNpy("/media/marcus/Vulcan/radiology/msh_hip/imgs/1104.15.1202.129.1155.118.20151106131219.124.npy")
hipsCohort() \%>\% `[[`(1, "img") \%>\% readImg() \%>\% Viz()
hipsCohort() \%>\% `[[`(1, "img") \%>\% showImg()
}
\concept{img}
