% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hips-package.R
\docType{data}
\name{scalars}
\alias{scalars}
\title{MSH cohort data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 23557 rows and 23 columns.}
\usage{
scalars
}
\description{
data.frame with 23,589 examples and 21 variables
}
\seealso{
\code{\link{hipsCohort}}

Other cohorts: \code{\link{adlCohort}},
  \code{\link{adl_test}}, \code{\link{caseControlCohorts}},
  \code{\link{hipsCohort_}}
}
\concept{cohorts}
\keyword{datasets}
