% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_aes.R
\name{scale_color_cnn}
\alias{scale_color_cnn}
\alias{aes_cnn_col}
\alias{scale_fill_cnn}
\title{Scales for CNNs}
\usage{
scale_color_cnn(title = "Encoding CNN", values = AES_CNN_COLORS,
  labels = AES_CNN_LABELS)

aes_cnn_col()

scale_fill_cnn(title = "Encoding CNN", values = AES_CNN_COLORS,
  labels = AES_CNN_LABELS)
}
\description{
Scales for CNNs
}
\examples{
vizR::pal_col(AES_CNN_COLORS)
ggplot(data.frame(x = 1:2, cnn = hips::hipsOpt(btlncks)), aes(x=x, fill=cnn)) +
  geom_bar() + scale_fill_cnn()
}
\seealso{
Other gg_aes: \code{\link{PrettyPredictors}},
  \code{\link{PrettyTargets}}, \code{\link{facet_cnn}},
  \code{\link{gg_x_target}},
  \code{\link{scale_color_cohort}},
  \code{\link{scale_color_device}},
  \code{\link{scale_color_fx}},
  \code{\link{scale_color_targetMode}},
  \code{\link{scale_color_target}},
  \code{\link{scale_color_view}},
  \code{\link{scale_fill_sex}},
  \code{\link{scale_predictor_}},
  \code{\link{scale_x_cohort}},
  \code{\link{scale_x_target}}
}
\concept{gg_aes}
