% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_aes.R
\name{scale_color_cohort}
\alias{scale_color_cohort}
\alias{aes_cohort_col}
\title{ggplot scale color by cohort}
\usage{
scale_color_cohort(title = "Test Cohort", values = AES_COHORT_COLORS,
  labels = AES_COHORT_LABELS)

aes_cohort_col()
}
\description{
ggplot scale color by cohort
}
\examples{
pal_col(AES_COHORT_COLORS)
}
\seealso{
Other gg_aes: \code{\link{PrettyPredictors}},
  \code{\link{PrettyTargets}}, \code{\link{facet_cnn}},
  \code{\link{gg_x_target}}, \code{\link{scale_color_cnn}},
  \code{\link{scale_color_device}},
  \code{\link{scale_color_fx}},
  \code{\link{scale_color_targetMode}},
  \code{\link{scale_color_target}},
  \code{\link{scale_color_view}},
  \code{\link{scale_fill_sex}},
  \code{\link{scale_predictor_}},
  \code{\link{scale_x_cohort}},
  \code{\link{scale_x_target}}
}
\concept{gg_aes}
