% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_aes.R
\name{scale_color_device}
\alias{scale_color_device}
\alias{aes_device_col}
\title{ggplot scale color by device}
\usage{
scale_color_device(title = "Scanner Model", values = AES_DEVICE_COLORS,
  labels = AES_DEVICE_LABELS)

aes_device_col()
}
\description{
ggplot scale color by device
}
\examples{
pal_col(AES_DEVICE_COLORS)
}
\seealso{
Other gg_aes: \code{\link{PrettyPredictors}},
  \code{\link{PrettyTargets}}, \code{\link{facet_cnn}},
  \code{\link{gg_x_target}}, \code{\link{scale_color_cnn}},
  \code{\link{scale_color_cohort}},
  \code{\link{scale_color_fx}},
  \code{\link{scale_color_targetMode}},
  \code{\link{scale_color_target}},
  \code{\link{scale_color_view}},
  \code{\link{scale_fill_sex}},
  \code{\link{scale_predictor_}},
  \code{\link{scale_x_cohort}},
  \code{\link{scale_x_target}}
}
\concept{gg_aes}
