% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_aes.R
\name{scale_color_targetMode}
\alias{scale_color_targetMode}
\title{Make a ggplot scale to color by target}
\usage{
scale_color_targetMode(title = "Target Type",
  labels = AES_TARGETMODE_LABELS, values = AES_TARGETMODE_COLORS)
}
\description{
Make a ggplot scale to color by target
}
\examples{
scale_color_target()
}
\seealso{
Other gg_aes: \code{\link{PrettyPredictors}},
  \code{\link{PrettyTargets}}, \code{\link{facet_cnn}},
  \code{\link{gg_x_target}}, \code{\link{scale_color_cnn}},
  \code{\link{scale_color_cohort}},
  \code{\link{scale_color_device}},
  \code{\link{scale_color_fx}},
  \code{\link{scale_color_target}},
  \code{\link{scale_color_view}},
  \code{\link{scale_fill_sex}},
  \code{\link{scale_predictor_}},
  \code{\link{scale_x_cohort}},
  \code{\link{scale_x_target}}
}
\concept{gg_aes}
