% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_aes.R
\name{scale_predictor_}
\alias{scale_predictor_}
\alias{scale_predictor}
\title{Scales for predictor sets.}
\usage{
scale_predictor_(aesthetic, title = "Explanatory Variables",
  values = AES_PREDICTOR_PALETTE, labels = AES_PREDICTOR_LABELS)

scale_predictor(aesthetic, ...)
}
\description{
Scales for predictor sets.
}
\examples{
vizR::pal_col(AES_PREDICTOR_PALETTE)
predictor_df <- data.frame(x = seq_along(AES_PREDICTOR_LABELS),
  data_value = names(AES_PREDICTOR_LABELS),
  pretty_value = unname(AES_PREDICTOR_LABELS))
ggplot(predictor_df, aes(x=x, fill=data_value)) +
  geom_bar() +
  scale_predictor_("fill")
ggplot(predictor_df, aes(x=x, col=data_value)) +
  geom_bar(fill="white") +
  scale_predictor_("color")
ggplot(predictor_df, aes(x=x, fill=data_value)) +
  geom_bar() +
  scale_predictor(fill)
ggplot(predictor_df, aes(x=x, col=data_value)) +
  geom_bar(fill="white") +
  scale_predictor(color)
}
\seealso{
Other gg_aes: \code{\link{PrettyPredictors}},
  \code{\link{PrettyTargets}}, \code{\link{facet_cnn}},
  \code{\link{gg_x_target}}, \code{\link{scale_color_cnn}},
  \code{\link{scale_color_cohort}},
  \code{\link{scale_color_device}},
  \code{\link{scale_color_fx}},
  \code{\link{scale_color_targetMode}},
  \code{\link{scale_color_target}},
  \code{\link{scale_color_view}},
  \code{\link{scale_fill_sex}},
  \code{\link{scale_x_cohort}},
  \code{\link{scale_x_target}}
}
\concept{gg_aes}
