% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_aes.R
\name{scale_x_target}
\alias{scale_x_target}
\alias{scale_y_target}
\title{ggplot scale to map targets to the x-axis}
\usage{
scale_x_target(title = "Classification Target",
  labels = AES_TARGET_LABELS, ...)

scale_y_target(title = "Classification Target",
  labels = AES_TARGET_LABELS, ...)
}
\description{
ggplot scale to map targets to the x-axis
}
\examples{
scale_x_target() \%>\% s
}
\seealso{
Other gg_aes: \code{\link{PrettyPredictors}},
  \code{\link{PrettyTargets}}, \code{\link{facet_cnn}},
  \code{\link{gg_x_target}}, \code{\link{scale_color_cnn}},
  \code{\link{scale_color_cohort}},
  \code{\link{scale_color_device}},
  \code{\link{scale_color_fx}},
  \code{\link{scale_color_targetMode}},
  \code{\link{scale_color_target}},
  \code{\link{scale_color_view}},
  \code{\link{scale_fill_sex}},
  \code{\link{scale_predictor_}},
  \code{\link{scale_x_cohort}}
}
\concept{gg_aes}
