% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_df.R
\name{StratifiedPartition}
\alias{StratifiedPartition}
\title{Augment cohort table with a stratified train/test partition}
\usage{
StratifiedPartition(cohort_df, grp_col = "pt", target_col = "fx")
}
\arguments{
\item{cohort_df}{data.frame}

\item{grp_col}{chr(1)}

\item{target_col}{chr(1)}
}
\value{
data.frame with same number of rows and an additional column `partition` with values train and test
}
\description{
this fxn wraps \code{\link[caret]{createDataPartition}}
}
\examples{
StratifiedPartition(mtcars, grp_col="am", target_col="mpg") \%>\% s
StratifiedPartition(mtcars, grp_col="am", target_col="mpg") \%>\% select(am, partition) \%>\% table_()
StratifiedPartition(hipsCohort(), grp_col = "device_model") \%>\% select(device_model, partition) \%>\% table_()
StratifiedPartition(hipsCohort()) \%>\% select(pt, partition) \%>\% table_()
try(StratifiedPartition(hipsCohort() \%>\% sample_n(500)) \%>\% select(pt, partition) \%>\% table_())
StratifiedPartition(hipsCohort()) \%>\% attributes() \%>\% s
}
