% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table.R
\name{Tbl}
\alias{Tbl}
\title{Save table file(s) for hips project}
\usage{
Tbl(x, bn = "", tbl_type = c("html", "latex", "xlsx"))
}
\arguments{
\item{x}{data.frame or kable table to be saved}

\item{bn}{chr(1) target file basename. defaults to name of `x`}

\item{tbl_type}{chr(n) types of table files to generate}
}
\value{
None. called for side effect
}
\description{
Save table file(s) for hips project
}
\examples{
Tbl(mtcars)
}
