% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eval.R
\name{train_eg_ids}
\alias{train_eg_ids}
\title{Extract vector of training examples used for a model}
\usage{
train_eg_ids(x)
}
\arguments{
\item{x}{caret model}
}
\value{
x_chr with identifiers for training data
}
\description{
Extract vector of training examples used for a model
}
\seealso{
Other evalTrain: \code{\link{predict_pY}},
  \code{\link{train_n_eg}}, \code{\link{train_terms}}
}
\concept{evalTrain}
