% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Train.R
\name{train_glm}
\alias{train_glm}
\alias{trainGlm}
\alias{trainTestGlm}
\title{Train a glm model}
\usage{
train_glm(train_df, target = "fx", predictors, id_col = "img")

trainGlm(train_df, target = "fx", predictor_set, id_col = "img")

trainTestGlm(cohort_df, target = "fx", predictor_set, grp_col = "pt",
  img_id = "img")
}
\arguments{
\item{train_df}{a radiograph-major data.frame with training examples.  See \code{StratifiedPartition} to partition data.}

\item{target}{chr(1) names of target to train}

\item{predictors}{x_chr names of predictor columns}

\item{btlnck}{}
}
\value{
a caret S3 train object
}
\description{
This function wraps \code{\link[caret]{train}} to train a glm model with my prefered settings and a
x_chr interface instead of a formula interface
}
\examples{
train_glm(hipsCohort() \%>\% unnest(pre), target = "fx", predictors = str_c("PC", 1:10))
trainGlm(hipsCohort(), predictor_set="image")
trainTestGlm(hipsCohort(), predictor_set = "image")
}
