% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Train.R
\name{train_lm}
\alias{train_lm}
\alias{trainLm}
\alias{trainTestLm}
\title{Base Linear Regression Model Function}
\usage{
train_lm(train_df, target, predictors)

trainLm(train_df, target, predictor_set)

trainTestLm(cohort_df, target, predictor_set, grp_col = "pt")
}
\arguments{
\item{train_df}{df with at least target and predictors cols}

\item{target}{chr(1)}

\item{predictors}{chr(n)}
}
\value{
caret train object
}
\description{
Base Linear Regression Model Function
}
\examples{
train_lm(hipsCohort(), target = "age", predictors = "device_model")
trainLm(hipsCohort(), target = "age", predictor_set = "image")
trainTestLm(hipsCohort(), target = "age", predictor_set = "image")
}
