% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eval.R
\name{train_n_eg}
\alias{train_n_eg}
\title{Extract total and limiting number of training examples from a caret model}
\usage{
train_n_eg(x)
}
\arguments{
\item{x}{caret model}
}
\value{
nmd_int(2) with n_limiting and n_train example counts
}
\description{
Extract total and limiting number of training examples from a caret model
}
\seealso{
Other evalTrain: \code{\link{predict_pY}},
  \code{\link{train_eg_ids}}, \code{\link{train_terms}}
}
\concept{evalTrain}
