% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eval.R
\name{train_terms}
\alias{train_terms}
\title{Extract predictor terms for a caret model}
\usage{
train_terms(x)
}
\arguments{
\item{x}{caret model}
}
\value{
x_chr of predictor terms. categoricals will be broken out to individual contrast levels
}
\description{
Extract predictor terms for a caret model
}
\seealso{
Other evalTrain: \code{\link{predict_pY}},
  \code{\link{train_eg_ids}}, \code{\link{train_n_eg}}
}
\concept{evalTrain}
