% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComplexSummary.R
\name{UniteCols}
\alias{UniteCols}
\title{Join compound table statistic columns.}
\usage{
UniteCols(x)
}
\arguments{
\item{x}{data.frame with individual statistics to join}
}
\description{
Join compound table statistic columns.
}
\examples{
map(grp_lst, compose(PrettyEach, BaseSummary), x=scalars)
map(grp_lst, compose(UniteCols, PrettyEach, BaseSummary), x=scalars)
map(grp_lst, compose(UniteCols, partial(PrettyEach, trim=TRUE), BaseSummary), x=scalars)
}
\seealso{
ComplexSummary
}
